# Optimize classnames enumeration