… modules/gdscript/gdscript.h
… modules/gdscript/tests/gdscript_test_runner.cpp
… modules/gdscript/tests/test_gdscript.cpp
… modules/mono/csharp_script.cpp
… modules/mono/csharp_script.h
… modules/mono/glue/GodotSharp/GodotSharp/Core/Bridge/ManagedCallbacks.cs
… modules/mono/glue/GodotSharp/GodotSharp/Core/Bridge/ScriptManagerBridge.cs
… modules/mono/mono_gd/gd_mono_cache.h