Hereby, a number of items involved in presenting icons of global classes are rewired to leverage the global class cache much better. As a consequence, the time it takes for the node create dialog and autocomplete popup to show up (especially in cold, just after opening the editor) is reduced considerably. This is especially impactful in big projects. I've been a witness of a moderately big project taking about 10 seconds for the node create dialog to be shown for it.

This PR also makes an initial commit that simplifies file system cache parsing, something one of the other commits leverages when it comes to including more attributes in the global class cache.

Benchmarking on a test project with thousands of trivial globally named GDScript and C# classes, some of them with icons, these are the results:
- Without this PR
	- Create node dialog
		Cold: 4692 ms (icon loads: 10, script loads: 2268)
		Hot: 1810 ms (icon loads: 0, script loads: 0)
	- Autocomplete popup
		Cold: 3079 ms (icon loads: 10, script loads: 756)
		Hot: 698 ms (icon loads: 0, script loads: 0)
- With this PR
	- Create node dialog
		Cold: 182 ms (icon loads: 10, script loads: 0
		Hot: 172 ms (icon loads: 0, script loads: 0)
	- Autocomplete popup
		Cold: 91 ms (icon loads: 10, script loads: 0)
		Hot: 83 ms (icon loads: 0, script loads: 0)

As the data above shows, one of the reasons for the speedup is that the new implementation doesn't need to load (potentially multiple times) all the scripts involved. The number of icon loads was measures as well to ensure no regressions in that area were caused by this PR. That is, the editor script icon cache is still used as most as possible.

Besides everything explained above, this PR removes part of the extra work the involved code was doing (such as checking the deprecated custom types' icons), to behave more consistently with what the file system dock does. Given the extent of the changes, there's some risk of regressions in picking up icons for classes.

Version of this PR for 4.3 submitted as #101490.

**UPDATE:** I forgot to add that another potential regression is that the create node dialog won't be able to show the global classes before they have been parsed and cached, which is especially true in the cold case. If that's true and a concern, we could still fall back to parsing the script files or let the create dialog wait until the global classes cache is ready.

- *Production edit: See https://github.com/godotengine/godot/issues/27333.*

Fixes #101194