@akien-mga 2025-01-13
CC @YYF233333 who has been working on #101435 recently, which seems partially related (both are probably relevant to merge but good to cross-review).