@YYF233333 2025-01-13
This one change the sign of `EditorNode::get_object_icon` as well as some change I don't come up with. It probably won't resolve RemoteTree performance issue though, because we are still doing `ResourceLoader::exists()` for every icon, so mine is still relevant. I think I can rebase on top of this once it is settled.