@KoBeWi 2025-01-22
Unfortunately this breaks custom icons for scripts that aren't global classes. `@icon` is supposed to work in any script, even built-in ones. It affects Scene dock.