@RandomShaper 2025-01-27
@KoBeWi, thanks for your detailed review. I'm glad you caught some issues and now I have a chance to fix them. I'll include fixes in my next push, including counterparts for the 4.3 version of the PR.