@RandomShaper 2025-01-28
OK. Pushed after addressing @KoBeWi's comments. I've kept most of the key function so the former algorithm can kick in unless the new fast path is able to fully leverage the class names cache.