@RandomShaper 2025-02-04
After rechecking the code with a fresher mind, I don't really think my concern was justified. What I'm changing in this PR is, in short, how the icon to a certain class name is obtained, once the class name is already known. In other words, if it's possible that you can get an incomplete type set in the create node dialog because the file system hasn't ended parsing the classes yet, that's without this PR as well.

However, given the first file system scan, which includes populating the inventory or global classes, happens synchronously on the main thread, and very early, so, unless I'm missing something, the user has no chance of interacting with the editor early enough for the issue to exist.