@akien-mga 2025-02-04
Tested the PR with Calinou's project and a dev build. It's extremely slow in `master`, and just pretty slow (5s) in this PR, but that's with an unoptimized build and 10000 classes, so that's expected. Calinou's test results show what an optimized production build yields.